requirements.txt
uvicorn
fastapi
pydantic
alpha_vantage
pandas
tensorflow
stripe
jinja2
python-dotenv